@@ -36,8 +36,6 @@
 #include "servers/audio/audio_rb_resampler.h"
 #include "servers/audio_server.h"
 
-class ImageTexture;
-
 class VideoStreamPlayer : public Control {
 	GDCLASS(VideoStreamPlayer, Control);
 
@@ -54,7 +52,7 @@ class VideoStreamPlayer : public Control {
 
 	RID stream_rid;
 
-	Ref<ImageTexture> texture;
+	Ref<Texture2D> texture;
 
 	AudioRBResampler resampler;
 	Vector<AudioFrame> mix_buffer;
